rename log type file to note
