+++ vue-rewrite/src/types/index.ts
@@ -1,5 +1,5 @@
 export * from "./day"
-export * from "./log"
+export * from "./note"
 export * from "./med"
 export * from "./meal"
 export * from "./symptom"
